extract woof
