+++ lib/handlers/moduleHandlers.js
@@ -1,4 +1,5 @@
 var tokenUtils = require('../util/tokenUtils');
+var functionHandlers = require('./functionHandlers');
 
 /**
  * Handles the so construct (thanks @maxogden!):
@@ -43,6 +44,70 @@ function handleSo(parseContext)
   return `var ${modName} = require('${lib}');\n`;
 }
 
+/**
+ * Handles the woof construct:
+ *  woof <alias> be <export>
+ *  woof <export>
+ *
+ * Produces:
+ *  module.exports.alias = export
+ *  module.exports = export
+ */
+function handleWoof(parseContext)
+{
+  tokenUtils.expectToken('woof', parseContext);
+
+  var tokens = parseContext.tokens;
+
+  // consume: woof
+  tokens.shift();
+
+  var exportName = '';
+
+  // look ahead and check if it is in `woof x be y` form
+  if ( tokens[1] === 'be' )
+  {
+    // woof foo be X -> module.exports.foo = X
+    exportName = '.' + tokens.shift();
+
+    // consume: be
+    tokens.shift();
+  }
+
+  // module.exports = SOMETHING
+  var statement = 'module.exports' + exportName + ' = ';
+
+  var assignmentValue = tokens[0];
+  // woof something -> module.exports = something
+  if( tokens.length === 1)
+  {
+    statement += tokens.shift();
+    statement += '\n';
+    return statement;
+  }
+
+  // module.exports = function x(a,b) {}
+  if ( assignmentValue === 'such' )
+  {
+    var functionStatement = functionHandlers.handleSuch(parseContext);
+    statement += functionStatement;
+    // the closing wow will be in a new line
+    return statement;
+  }
+
+  // module.exports = function (a,b) {}
+  if ( assignmentValue === 'much')
+  {
+    var anonStatement = functionHandlers.handleLambda(parseContext);
+    statement += anonStatement;
+    // the closing wow will be in a new line
+    return statement;
+  }
+
+  // TODO support other expressions
+  throw new SyntaxError(`Invalid parse state! Expected a value but got: '${tokens[0]}' from chain: [${tokens}]. Allowed construct 'woof [<name> be] <value | <SUCH> | <MUCH> >'. ${tokenUtils.parseInfo(parseContext)}`);
+}
+
 module.exports = {
-    handleSo
+    handleSo, handleWoof
 }
